Uses new syntax to support a swift way to handle errors and exceptions

+++ Pod/Classes/CIRResultSet.m
@@ -57,7 +57,7 @@
 	return [_statement isClosed];
 }
 
-- (int)columnCount
+- (NSUInteger)columnCount
 {
 	return [_statement columnCount];
 }
